backend/requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-JWT-Extended==4.6.0
Flask-Cors==4.0.1
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1
pandas==2.2.2
openpyxl==3.1.5
psycopg2-binary==2.9.9
gunicorn==22.0.0
